fix: uaa route-registrar had wrong key in config
There was a typo in the route-registrar configuration (service ->
server), which caused it to not work correctly (as the TLS certificate
could not provide the host name).

Also move more scripts into separate files; eventually we should bake
them into an init image or something, but currently we have no
infrastructure to create images.